backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
pandas
openpyxl
python-dotenv
